finish

+++ ParkingSystem/MasterMember.cs
@@ -26,6 +26,7 @@ namespace ParkingSystem
         private DataContext? _context;
         private string? operation;
         private int? selectedMemberId;
+        private string? type;
 
         public MasterMember()
         {
@@ -54,7 +55,7 @@ namespace ParkingSystem
         {
             base.OnClosed(e);
             _context.Dispose();
-            _context= null;
+            _context = null;
         }
 
         private void EnabledTrue()
@@ -120,6 +121,7 @@ namespace ParkingSystem
         {
             EnabledTrue();
             SelectMemType();
+            cmbMemberType.Text = type;
 
             operation = "update";
         }
@@ -128,18 +130,18 @@ namespace ParkingSystem
         {
             switch (operation)
             {
-                case "insert" : 
-                        actionInsert();
+                case "insert":
+                    actionInsert();
                     break;
                 case "update":
-                        actionUpdate();
+                    actionUpdate();
                     break;
                 case "delete":
-                        actionDelete();
+                    actionDelete();
                     break;
                 default: break;
             }
-             dataGridView1.Refresh();
+            dataGridView1.Refresh();
         }
 
         private async void actionInsert()
@@ -150,12 +152,13 @@ namespace ParkingSystem
 
             if (!cPhoneNum || !cDateBirth)
             {
-                MessageBox.Show("Data yang Anda Masukkan tidak sesuai");
-            } 
-            else if (txtName.Text == "" || txtEmail.Text == "" || RTAddress.Text == "" || cmbMemberType.Text == "" )
+                MessageBox.Show("The data you entered is not correct", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            else if (txtName.Text == "" || txtEmail.Text == "" || RTAddress.Text == "" || cmbMemberType.Text == "")
             {
-                MessageBox.Show("Anda harus memasukkan data dengan lengkap");
-            } else
+                MessageBox.Show("You must enter data completely", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            else
             {
                 // abis itu get member type
                 string textTypeid = cmbMemberType.SelectedItem.ToString();
@@ -176,18 +179,21 @@ namespace ParkingSystem
                 {
                     newMember.Gender = "Male";
                 }
-                else if (radFemale.Checked) 
+                else if (radFemale.Checked)
                 {
                     newMember.Gender = "Female";
-                } else
-                {
-                    MessageBox.Show("Anda harus menginput gender");
                 }
+                else
+                {
+                    MessageBox.Show("You must input gender", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
+                newMember.CreatedAt = DateTime.Now;
 
                 _context.Members.Add(newMember);
                 await _context.SaveChangesAsync();
                 _context.Members.Load();
                 dataGridView1.Refresh();
+                MessageBox.Show("Successfully inserted new member data with ID : " + newMember.Id, "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
             }
 
         }
@@ -197,13 +203,21 @@ namespace ParkingSystem
             // cek selected member id nya ada atau tidak, kalau tidak diem aja
             if (selectedMemberId != null)
             {
-                var member = await _context.Members.Where(m => m.Id == selectedMemberId).FirstAsync();
-                _context.Members.Remove(member);
-                await _context.SaveChangesAsync();
-                selectedMemberId = null;
-                _context.Members.Load();
-                dataGridView1.Refresh();
-            } else
+                if (MessageBox.Show("Are you sure you want to delete this data?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.No)
+                {
+                    return;
+                }
+                else
+                {
+                    var member = await _context.Members.Where(m => m.Id == selectedMemberId).FirstAsync();
+                    _context.Members.Remove(member);
+                    await _context.SaveChangesAsync();
+                    selectedMemberId = null;
+                    _context.Members.Load();
+                    dataGridView1.Refresh();
+                }
+            }
+            else
             {
                 MessageBox.Show("Member yang ingin anda hapus tidak ada dalam database");
             }
@@ -222,7 +236,7 @@ namespace ParkingSystem
 
                 // get member dan simpan ke db
                 var member = await _context.Members
-                                .Where(m => m.Id == selectedMemberId) 
+                                .Where(m => m.Id == selectedMemberId)
                                 .FirstAsync();
                 member.Name = txtName.Text;
                 member.Type = memType;
@@ -238,13 +252,17 @@ namespace ParkingSystem
                 {
                     member.Gender = "Female";
                 }
+                member.LastUpdatedAt = DateTime.Now;
+
                 _context.Members.Update(member);
                 await _context.SaveChangesAsync();
                 _context.Members.Load();
                 dataGridView1.Refresh();
-            } else
+                MessageBox.Show("Successfully updated new vehicle data with ID : " + member.Id, "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            }
+            else
             {
-                MessageBox.Show("Member yang ingin anda update tidak ada dalam database");
+                MessageBox.Show("the data does not exist in the database", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
         }
         private void dataGridView1_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -262,12 +280,28 @@ namespace ParkingSystem
             if (row.Cells[8].Value.ToString() == "Male")
             {
                 radMale.Checked = true;
-            } else 
+            }
+            else
             {
                 radFemale.Checked = true;
             }
-            string type = row.Cells[0].Value.ToString();
-            cmbMemberType.Text= type;
+            type = row.Cells[0].Value.ToString();
+            cmbMemberType.Text = type;
+
+            var created = row.Cells[9].Value;
+            var updated = row.Cells[10].Value;
+            if (updated != null)
+            {
+                lblModifed.Text = "This record is last modifed at " + updated;
+            }
+            else if (created != null)
+            {
+                lblModifed.Text = "This record is created at " + created;
+            }
+            else if (updated == null && created == null)
+            {
+                lblModifed.Text = "";
+            }
 
             txtName.Enabled = false;
             txtEmail.Enabled = false;
